Provide go.env if missing

+++ rpm/go1.21.spec
@@ -106,6 +106,8 @@ URL:            https://github.com/sailfishos-mirror/go/
 Source:         %{name}-%{version}.tar.xz
 Source1:        go-rpmlintrc
 Source6:        go.gdbinit
+# For cases where go.env isn't available/created. From Go 1.21 binary rpm.
+Source7:        go.env
 # We have to compile TSAN ourselves. boo#1052528
 # Preferred form when all arches share llvm race version
 # Source100:      llvm-%%{tsan_commit}.tar.xz
@@ -321,7 +323,11 @@ install -d  %{buildroot}%{_datadir}/go/%{go_label}/contrib/src
 ln -s %{_datadir}/go/%{go_label}/contrib/src/ %{buildroot}%{_libdir}/go/%{go_label}/contrib/src
 
 # go.env sets defaults for: GOPROXY GOSUMDB GOTOOLCHAIN
-install -Dm644 go.env $GOROOT/
+if [ -f go.env ]; then
+  install -Dm644 go.env $GOROOT/
+else
+  install -Dm644 %{SOURCE7} $GOROOT/
+fi
 
 # Change go.env GOTOOLCHAIN default to "local" so Go app builds never
 # automatically download newer toolchains as specified by go.mod files.
